requirements.txt
discord.py==2.2.2
python-dotenv==1.0.0
PyYAML==6.0
supabase==1.0.3
